Impacts # 29 & # 33:
1) This fixes a backwards compatibility issue with # 29 for politicians
that have no county or city defined.
2) For issue 33 the order of edit and add was swapped.  Also, edit now
requires a user to have at least two points.  The reason behind this
change is that OG Whiz went into Edit when she meant to add a rating.
First time users should probably just have the opportunity to edit so
that they aren't overwhelmed or click the wrong thing.